fix(task-runner): cast RETURNING 1 to BIGINT for sqlx i64 compatibility
PostgreSQL literal '1' returns INT4, but query_scalar::<_, i64> expects
INT8. This caused 'mismatched types' runtime errors in task executors.

+++ crates/task-runner/src/executor/activities.rs
@@ -209,7 +209,7 @@ async fn rebuild_activities_batch(pool: &PgPool, start_block: i64, end_block: i6
             amount, asset_id, metadata, timestamp
         )
         SELECT * FROM all_activities
-        RETURNING 1
+        RETURNING 1::BIGINT
         "#,
     )
     .bind(start_block)
